requirements.txt
Pillow>=10.0.0
pystray>=0.19.0
requests>=2.31.0
schedule>=1.2.0
